casting labels to long before passing to torchmetrics

+++ base.py
@@ -42,7 +42,7 @@ class BaseModel(L.LightningModule):
         output = {"labels": y, "probabilities": probabilities, "loss": batch_loss}
         self.training_step_outputs.append(output)
 
-        self.metric_collection.update(probabilities, y)
+        self.metric_collection.update(probabilities, y.long()) # Cast to long type for metrics
 
         return batch_loss #what we return is irrelevant in latest lightning version
 
@@ -62,7 +62,7 @@ class BaseModel(L.LightningModule):
         output = {"labels": y, "probabilities": probabilities, "loss": batch_loss}
         self.validation_step_outputs.append(output)
 
-        self.metric_collection.update(probabilities, y)
+        self.metric_collection.update(probabilities, y.long()) # Cast to long type for metrics
 
         return output
 
@@ -82,7 +82,7 @@ class BaseModel(L.LightningModule):
         output = {"labels": y, "probabilities": probabilities, "loss": batch_loss}
         self.test_step_outputs.append(output)
 
-        self.metric_collection.update(probabilities, y)
+        self.metric_collection.update(probabilities, y.long()) # Cast to long type for metrics
 
         return output
